// dashboard/src/api/models/runSnapshotEvent.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Scheduler Public API (v1)
 * OpenAPI spec version: 1.3.0
 */
import type { Run } from './run';
import type { RunNodeStatus } from './runNodeStatus';

export interface RunSnapshotEvent {
  kind: 'run.snapshot';
  run: Run;
  /** @nullable */
  nodes?: RunNodeStatus[] | null;
}